add field is_default in category, modify categoryRepository and entity and add fix connection

+++ src/main/java/university/jala/finalProject/springJPA/service/CategoryServiceImpl.java
@@ -59,9 +59,12 @@ public class CategoryServiceImpl implements CategoryService {
         if (listCount > 0) {
             throw new RuntimeException("Cannot delete, category has " + listCount + " associated lists!");
         }
-
-        categoryRepository.delete(category);
-        return true;
+        if( category.getDefault() == true) {
+            return false;
+        } else {
+            categoryRepository.delete(category);
+            return true;
+        }
     }
 
     @Override
